pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name="Swikrit Shukla", email="work.swikrit@gmail.com"}]
license = "MIT"                     # <-- SPDX short-expression (string)
license-files = ["LICENSE"]   
dependencies = [
  "google-genai>=1.33.0",
  "python-dotenv>=1.1.1",
  "streamlit>=1.51.0",
]

[project.scripts]
coding-agent = "coding_agent.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["coding_agent*"]
# optionally exclude dummy if you must
exclude = ["dummy*","dummy"]

